create a log driver system, add loki driver, add prometheus metrics, update docs
