+++ .gitignore
@@ -8,3 +8,9 @@ target/
 
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
+
+# nix
+.direnv
+result*
+
+.vscode
