+++ Cargo.toml
@@ -9,12 +9,15 @@ needless_return = "allow"
 
 [dependencies]
 anyhow = "1.0.79"
+async-trait = "0.1.81"
 aws-config = "1.1.8"
 aws-sdk-cloudwatchlogs = "1.16.0"
 axum = { version = "0.7.4", features = ["tracing"] }
 axum-extra = { version = "0.9.2", features = ["typed-header"] }
+axum-prometheus = "0.7.0"
 clap = { version = "4.4.18", features = ["derive", "env"] }
 hex = "0.4.3"
+reqwest = { version = "0.12.5", features = ["json"] }
 ring = "0.17.7"
 serde = { version = "1.0.196", features = ["derive"] }
 serde_json = "1.0.112"
