+++ src/main.rs
@@ -1,13 +1,19 @@
 mod controller;
+mod drivers;
 mod types;
 
+use crate::drivers::{CloudWatchDriver, LokiDriver};
+use crate::types::LogDriver;
 use anyhow::Result;
 use axum::{
     body::{Body, Bytes},
     extract::State,
     http::{header::HeaderMap, Response, StatusCode},
     response::IntoResponse,
+    routing::get,
 };
+use axum_prometheus::metrics::counter;
+use axum_prometheus::PrometheusMetricLayerBuilder;
 use clap::Parser;
 use ring::hmac;
 use tokio::signal::{unix, unix::SignalKind};
@@ -28,6 +34,23 @@ struct Args {
     vercel_verify: String,
     #[arg(long, env = "VERCEL_SECRET")]
     vercel_secret: String,
+
+    #[arg(long, env = "VERCEL_LOG_DRAIN_ENABLE_METRICS")]
+    enable_metrics: bool,
+    #[arg(long, env = "VERCEL_LOG_DRAIN_METRICS_PREFIX", default_value = "drain")]
+    metrics_prefix: String,
+
+    #[arg(long, env = "VERCEL_LOG_DRAIN_ENABLE_CLOUDWATCH")]
+    enable_cloudwatch: bool,
+
+    #[arg(long, env = "VERCEL_LOG_DRAIN_ENABLE_LOKI")]
+    enable_loki: bool,
+    #[arg(long, env = "VERCEL_LOG_DRAIN_LOKI_URL", default_value = "")]
+    loki_url: String,
+    #[arg(long, env = "VERCEL_LOG_DRAIN_LOKI_USER", default_value = "")]
+    loki_basic_auth_user: String,
+    #[arg(long, env = "VERCEL_LOG_DRAIN_LOKI_PASS", default_value = "")]
+    loki_basic_auth_pass: String,
 }
 
 #[derive(Debug, Clone)]
@@ -45,18 +68,33 @@ async fn main() -> Result<()> {
         .with_max_level(args.log)
         .init();
 
-    let config = aws_config::load_defaults(aws_config::BehaviorVersion::v2023_11_09()).await;
-
-    let cwl_client = aws_sdk_cloudwatchlogs::Client::new(&config);
     let (tx, rx) = mpsc::unbounded_channel::<types::Message>();
-    let mut controller = controller::Controller::new(tx.clone(), rx, cwl_client);
 
-    controller.init_aws_state().await?;
+    let mut drivers: Vec<Box<dyn LogDriver>> = Vec::new();
+
+    if args.enable_cloudwatch {
+        let config = aws_config::load_defaults(aws_config::BehaviorVersion::v2023_11_09()).await;
+        let cwl_client = aws_sdk_cloudwatchlogs::Client::new(&config);
+        drivers.push(Box::new(CloudWatchDriver::new(cwl_client)));
+        debug!("added cloudwatch driver");
+    }
+
+    if args.enable_loki {
+        drivers.push(Box::new(LokiDriver::new(
+            args.loki_url,
+            args.loki_basic_auth_user,
+            args.loki_basic_auth_pass,
+        )));
+        debug!("added loki driver");
+    }
+
+    let mut controller = controller::Controller::new(tx.clone(), rx, drivers);
+
+    controller.init().await?;
 
     tokio::spawn(async move {
         controller.run().await;
     });
-
     let state = AppState {
         vercel_verify: args.vercel_verify,
         vercel_secret: hmac::Key::new(
@@ -69,11 +107,21 @@ async fn main() -> Result<()> {
     let listen_address = format!("{}:{}", args.ip, args.port);
     let listener = tokio::net::TcpListener::bind(listen_address.clone()).await?;
 
-    let app = axum::Router::new()
+    let mut app = axum::Router::new()
         .route("/", axum::routing::post(root))
         .route("/health", axum::routing::get(health_check))
         .route("/vercel", axum::routing::post(ingest))
         .with_state(state);
+
+    if args.enable_metrics {
+        let (prometheus_layer, metric_handle) = PrometheusMetricLayerBuilder::new()
+            .with_prefix(args.metrics_prefix)
+            .with_default_metrics()
+            .build_pair();
+        app = app
+            .route("/metrics", get(|| async move { metric_handle.render() }))
+            .layer(prometheus_layer);
+    }
 
     info!("Listening on {}", listen_address);
     axum::serve(
@@ -132,6 +180,7 @@ async fn ingest(
         Some(signature) => signature.to_str().unwrap(),
         None => {
             warn!("received payload without signature");
+            counter!("drain_recv_invalid_signature").increment(1);
             return response;
         }
     };
@@ -139,6 +188,7 @@ async fn ingest(
         Ok(body_string) => body_string,
         Err(e) => {
             error!("received bad utf-8: {:?}", e);
+            counter!("drain_recv_bad_utf8").increment(1);
             return response;
         }
     };
@@ -148,6 +198,7 @@ async fn ingest(
         Ok(_) => {}
         Err(e) => {
             error!("failed verifying signature: {:?}", e);
+            counter!("drain_failed_verify_signature").increment(1);
             return response;
         }
     }
@@ -158,7 +209,7 @@ async fn ingest(
                 match state.log_queue.send(message) {
                     Ok(_) => {}
                     Err(e) => {
-                        error!("failed to queue log message to be sent to aws: {:?}", e);
+                        error!("failed to queue log message to be sent to outputs: {:?}", e);
                     }
                 }
             }
